no changes

+++ contracts/src/Main.sol
@@ -4,14 +4,14 @@ pragma solidity ^0.8;
 import "./Collection.sol";
 
 contract Main {
-  int private count;
-  mapping(int => Collection) private collections;
+  uint private count;
+  mapping(uint => Collection) private collections;
 
   constructor() {
     count = 0;
   }
 
-  function createCollection(string calldata name, int cardCount) external {
+  function createCollection(string calldata name, uint256 cardCount) external {
     collections[count++] = new Collection(name, cardCount);
   }
 }
